Change everything to embrace mutable dicts
Going into this, I had in mind that I would try to use a fully
functional approach, including the use of immutable datastructures.

As a quick hack, I just applied `deepcopy` to everything whenever I
needed to return a new dict, but I was still modifying in place within
the functions.

This felt off, so I decided to just get rid of the idea altogether.
Dictionaries are mutable in Python.