Remove unnecessary API endpoints and change PUT
The big change here is that PUT /characters/<id> now runs
`rerole_lib.character.calculate()` on the data, and returns the data
itself, rather than a link to it. This greatly simplifies the usage of
the API, and allows us to remove the `/calculate` endpoint.

In addition, due to the internal changes of how antimagic fields are
included in the calculation, we can remove the antimagic field
endpoints, as they are no longer necessary.