Add POC "Effect" selector and demo of functionality
In order to interact with things like feats, traits, etc, we need to be
able to add effects to them. This new component is a basic interface for
specifying effect details, and can return a bare "effect" object for use
throughout the character sheet.

Lots of refactoring etc. to do, but for now, this is a relatively stable
POC implementation.