Move main 'character' functionality into new Sheet class
The intention here is to make this new `Sheet` class the main top-level
export of the rerole-lib module. It has a narrow API, and can be
imported right out of the module:

    from rerole_lib import Sheet

Some of the previous functionality in the character module has been
condensed, made redundant, and/or renamed as part of the move.

This is a follow-up to the previous change that made a new `Dict` class
in `utils`.