Update effect index building and antimagic field function to work with sources
Instead of working with raw "effects," the functions that deal with
effects at the character sheet level now deal with effect sources.

Effects are now contained in a collection within an effect source, under
the "effects" key.

Effect sources can be active or inactive, magical or non-magical; when
the index is built, it checks for active effect sources, instead of all
effects. Effect state (active, inactive, etc.) is now checked at the
source, instead of within each effect.

This means that the logic for effect activity can be removed from
effect.applied().